\name{phoneme}
\alias{phoneme}
\docType{data}
\title{ Data From a Acoustic-Phonetic Continuous Speech Corpus}
\description{
  See Details.
}
\usage{data(phoneme)}
\format{
  A data frame with 4509 observations on the following 258 variables.
  \describe{
    \item{x.1}{a numeric vector}
    \item{x.2}{a numeric vector}
    \item{x.3}{a numeric vector}
    \item{x.4}{a numeric vector}
    \item{x.5}{a numeric vector}
    \item{x.6}{a numeric vector}
    \item{x.7}{a numeric vector}
    \item{x.8}{a numeric vector}
    \item{x.9}{a numeric vector}
    \item{x.10}{a numeric vector}
    \item{x.11}{a numeric vector}
    \item{x.12}{a numeric vector}
    \item{x.13}{a numeric vector}
    \item{x.14}{a numeric vector}
    \item{x.15}{a numeric vector}
    \item{x.16}{a numeric vector}
    \item{x.17}{a numeric vector}
    \item{x.18}{a numeric vector}
    \item{x.19}{a numeric vector}
    \item{x.20}{a numeric vector}
    \item{x.21}{a numeric vector}
    \item{x.22}{a numeric vector}
    \item{x.23}{a numeric vector}
    \item{x.24}{a numeric vector}
    \item{x.25}{a numeric vector}
    \item{x.26}{a numeric vector}
    \item{x.27}{a numeric vector}
    \item{x.28}{a numeric vector}
    \item{x.29}{a numeric vector}
    \item{x.30}{a numeric vector}
    \item{x.31}{a numeric vector}
    \item{x.32}{a numeric vector}
    \item{x.33}{a numeric vector}
    \item{x.34}{a numeric vector}
    \item{x.35}{a numeric vector}
    \item{x.36}{a numeric vector}
    \item{x.37}{a numeric vector}
    \item{x.38}{a numeric vector}
    \item{x.39}{a numeric vector}
    \item{x.40}{a numeric vector}
    \item{x.41}{a numeric vector}
    \item{x.42}{a numeric vector}
    \item{x.43}{a numeric vector}
    \item{x.44}{a numeric vector}
    \item{x.45}{a numeric vector}
    \item{x.46}{a numeric vector}
    \item{x.47}{a numeric vector}
    \item{x.48}{a numeric vector}
    \item{x.49}{a numeric vector}
    \item{x.50}{a numeric vector}
    \item{x.51}{a numeric vector}
    \item{x.52}{a numeric vector}
    \item{x.53}{a numeric vector}
    \item{x.54}{a numeric vector}
    \item{x.55}{a numeric vector}
    \item{x.56}{a numeric vector}
    \item{x.57}{a numeric vector}
    \item{x.58}{a numeric vector}
    \item{x.59}{a numeric vector}
    \item{x.60}{a numeric vector}
    \item{x.61}{a numeric vector}
    \item{x.62}{a numeric vector}
    \item{x.63}{a numeric vector}
    \item{x.64}{a numeric vector}
    \item{x.65}{a numeric vector}
    \item{x.66}{a numeric vector}
    \item{x.67}{a numeric vector}
    \item{x.68}{a numeric vector}
    \item{x.69}{a numeric vector}
    \item{x.70}{a numeric vector}
    \item{x.71}{a numeric vector}
    \item{x.72}{a numeric vector}
    \item{x.73}{a numeric vector}
    \item{x.74}{a numeric vector}
    \item{x.75}{a numeric vector}
    \item{x.76}{a numeric vector}
    \item{x.77}{a numeric vector}
    \item{x.78}{a numeric vector}
    \item{x.79}{a numeric vector}
    \item{x.80}{a numeric vector}
    \item{x.81}{a numeric vector}
    \item{x.82}{a numeric vector}
    \item{x.83}{a numeric vector}
    \item{x.84}{a numeric vector}
    \item{x.85}{a numeric vector}
    \item{x.86}{a numeric vector}
    \item{x.87}{a numeric vector}
    \item{x.88}{a numeric vector}
    \item{x.89}{a numeric vector}
    \item{x.90}{a numeric vector}
    \item{x.91}{a numeric vector}
    \item{x.92}{a numeric vector}
    \item{x.93}{a numeric vector}
    \item{x.94}{a numeric vector}
    \item{x.95}{a numeric vector}
    \item{x.96}{a numeric vector}
    \item{x.97}{a numeric vector}
    \item{x.98}{a numeric vector}
    \item{x.99}{a numeric vector}
    \item{x.100}{a numeric vector}
    \item{x.101}{a numeric vector}
    \item{x.102}{a numeric vector}
    \item{x.103}{a numeric vector}
    \item{x.104}{a numeric vector}
    \item{x.105}{a numeric vector}
    \item{x.106}{a numeric vector}
    \item{x.107}{a numeric vector}
    \item{x.108}{a numeric vector}
    \item{x.109}{a numeric vector}
    \item{x.110}{a numeric vector}
    \item{x.111}{a numeric vector}
    \item{x.112}{a numeric vector}
    \item{x.113}{a numeric vector}
    \item{x.114}{a numeric vector}
    \item{x.115}{a numeric vector}
    \item{x.116}{a numeric vector}
    \item{x.117}{a numeric vector}
    \item{x.118}{a numeric vector}
    \item{x.119}{a numeric vector}
    \item{x.120}{a numeric vector}
    \item{x.121}{a numeric vector}
    \item{x.122}{a numeric vector}
    \item{x.123}{a numeric vector}
    \item{x.124}{a numeric vector}
    \item{x.125}{a numeric vector}
    \item{x.126}{a numeric vector}
    \item{x.127}{a numeric vector}
    \item{x.128}{a numeric vector}
    \item{x.129}{a numeric vector}
    \item{x.130}{a numeric vector}
    \item{x.131}{a numeric vector}
    \item{x.132}{a numeric vector}
    \item{x.133}{a numeric vector}
    \item{x.134}{a numeric vector}
    \item{x.135}{a numeric vector}
    \item{x.136}{a numeric vector}
    \item{x.137}{a numeric vector}
    \item{x.138}{a numeric vector}
    \item{x.139}{a numeric vector}
    \item{x.140}{a numeric vector}
    \item{x.141}{a numeric vector}
    \item{x.142}{a numeric vector}
    \item{x.143}{a numeric vector}
    \item{x.144}{a numeric vector}
    \item{x.145}{a numeric vector}
    \item{x.146}{a numeric vector}
    \item{x.147}{a numeric vector}
    \item{x.148}{a numeric vector}
    \item{x.149}{a numeric vector}
    \item{x.150}{a numeric vector}
    \item{x.151}{a numeric vector}
    \item{x.152}{a numeric vector}
    \item{x.153}{a numeric vector}
    \item{x.154}{a numeric vector}
    \item{x.155}{a numeric vector}
    \item{x.156}{a numeric vector}
    \item{x.157}{a numeric vector}
    \item{x.158}{a numeric vector}
    \item{x.159}{a numeric vector}
    \item{x.160}{a numeric vector}
    \item{x.161}{a numeric vector}
    \item{x.162}{a numeric vector}
    \item{x.163}{a numeric vector}
    \item{x.164}{a numeric vector}
    \item{x.165}{a numeric vector}
    \item{x.166}{a numeric vector}
    \item{x.167}{a numeric vector}
    \item{x.168}{a numeric vector}
    \item{x.169}{a numeric vector}
    \item{x.170}{a numeric vector}
    \item{x.171}{a numeric vector}
    \item{x.172}{a numeric vector}
    \item{x.173}{a numeric vector}
    \item{x.174}{a numeric vector}
    \item{x.175}{a numeric vector}
    \item{x.176}{a numeric vector}
    \item{x.177}{a numeric vector}
    \item{x.178}{a numeric vector}
    \item{x.179}{a numeric vector}
    \item{x.180}{a numeric vector}
    \item{x.181}{a numeric vector}
    \item{x.182}{a numeric vector}
    \item{x.183}{a numeric vector}
    \item{x.184}{a numeric vector}
    \item{x.185}{a numeric vector}
    \item{x.186}{a numeric vector}
    \item{x.187}{a numeric vector}
    \item{x.188}{a numeric vector}
    \item{x.189}{a numeric vector}
    \item{x.190}{a numeric vector}
    \item{x.191}{a numeric vector}
    \item{x.192}{a numeric vector}
    \item{x.193}{a numeric vector}
    \item{x.194}{a numeric vector}
    \item{x.195}{a numeric vector}
    \item{x.196}{a numeric vector}
    \item{x.197}{a numeric vector}
    \item{x.198}{a numeric vector}
    \item{x.199}{a numeric vector}
    \item{x.200}{a numeric vector}
    \item{x.201}{a numeric vector}
    \item{x.202}{a numeric vector}
    \item{x.203}{a numeric vector}
    \item{x.204}{a numeric vector}
    \item{x.205}{a numeric vector}
    \item{x.206}{a numeric vector}
    \item{x.207}{a numeric vector}
    \item{x.208}{a numeric vector}
    \item{x.209}{a numeric vector}
    \item{x.210}{a numeric vector}
    \item{x.211}{a numeric vector}
    \item{x.212}{a numeric vector}
    \item{x.213}{a numeric vector}
    \item{x.214}{a numeric vector}
    \item{x.215}{a numeric vector}
    \item{x.216}{a numeric vector}
    \item{x.217}{a numeric vector}
    \item{x.218}{a numeric vector}
    \item{x.219}{a numeric vector}
    \item{x.220}{a numeric vector}
    \item{x.221}{a numeric vector}
    \item{x.222}{a numeric vector}
    \item{x.223}{a numeric vector}
    \item{x.224}{a numeric vector}
    \item{x.225}{a numeric vector}
    \item{x.226}{a numeric vector}
    \item{x.227}{a numeric vector}
    \item{x.228}{a numeric vector}
    \item{x.229}{a numeric vector}
    \item{x.230}{a numeric vector}
    \item{x.231}{a numeric vector}
    \item{x.232}{a numeric vector}
    \item{x.233}{a numeric vector}
    \item{x.234}{a numeric vector}
    \item{x.235}{a numeric vector}
    \item{x.236}{a numeric vector}
    \item{x.237}{a numeric vector}
    \item{x.238}{a numeric vector}
    \item{x.239}{a numeric vector}
    \item{x.240}{a numeric vector}
    \item{x.241}{a numeric vector}
    \item{x.242}{a numeric vector}
    \item{x.243}{a numeric vector}
    \item{x.244}{a numeric vector}
    \item{x.245}{a numeric vector}
    \item{x.246}{a numeric vector}
    \item{x.247}{a numeric vector}
    \item{x.248}{a numeric vector}
    \item{x.249}{a numeric vector}
    \item{x.250}{a numeric vector}
    \item{x.251}{a numeric vector}
    \item{x.252}{a numeric vector}
    \item{x.253}{a numeric vector}
    \item{x.254}{a numeric vector}
    \item{x.255}{a numeric vector}
    \item{x.256}{a numeric vector}
    \item{g}{a factor with levels \code{aa} 
              \code{ao} \code{dcl} \code{iy} \code{sh}}
    \item{speaker}{a factor with 437 levels}
  }
}
\details{
  These data arose from a collaboration between  Andreas Buja, Werner
  Stuetzle and Martin Maechler, and we used as an illustration in the
  paper on Penalized Discriminant Analysis by Hastie, Buja and
  Tibshirani (1995), referenced in the text. 

  The data were extracted from the TIMIT database (TIMIT
  Acoustic-Phonetic Continuous Speech Corpus, NTIS, US Dept of Commerce)
  which is a widely used resource for research in speech recognition.  A
  dataset was formed by selecting five phonemes for
  classification based on digitized speech from this database.  The
  phonemes are transcribed as follows: "sh" as in "she", "dcl" as in
  "dark", "iy" as the vowel in "she", "aa" as the vowel in "dark", and
  "ao" as the first vowel in "water".  From continuous speech of 50 male
  speakers, 4509 speech frames of 32 msec duration were selected,
  approximately 2 examples of each phoneme from each speaker.  Each
  speech frame is represented by 512 samples at a 16kHz sampling rate,
  and each frame represents one of the above five phonemes.  The
  breakdown of the 4509 speech frames into phoneme frequencies is as
  follows:

  aa   ao dcl   iy  sh 
 695 1022 757 1163 872

  From each speech frame, we computed a log-periodogram, which is one of
  several widely used methods for casting speech data in a form suitable
  for speech recognition.  Thus the data used in what follows consist of
  4509 log-periodograms of length 256, with known class (phoneme)
  memberships. 

  The data contain 256 columns labelled "x.1" - "x.256", a response
  column labelled "g", and a column labelled "speaker" identifying the
  different speakers.
}
\examples{
 head(str(phoneme))
}
\keyword{datasets}
