\name{waveform.test}
\alias{waveform.test}
\docType{data}
\title{ Simulated Waveform Data  }
\description{
   This is a simulated three-class problem with 21 variables, and 
   considered to be a difficult pattern recognition problem. 
}
\usage{data(waveform.test)}
\format{
  A data frame with 500 observations on the following 22 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x.1}}{a numeric vector}
    \item{\code{x.2}}{a numeric vector}
    \item{\code{x.3}}{a numeric vector}
    \item{\code{x.4}}{a numeric vector}
    \item{\code{x.5}}{a numeric vector}
    \item{\code{x.6}}{a numeric vector}
    \item{\code{x.7}}{a numeric vector}
    \item{\code{x.8}}{a numeric vector}
    \item{\code{x.9}}{a numeric vector}
    \item{\code{x.10}}{a numeric vector}
    \item{\code{x.11}}{a numeric vector}
    \item{\code{x.12}}{a numeric vector}
    \item{\code{x.13}}{a numeric vector}
    \item{\code{x.14}}{a numeric vector}
    \item{\code{x.15}}{a numeric vector}
    \item{\code{x.16}}{a numeric vector}
    \item{\code{x.17}}{a numeric vector}
    \item{\code{x.18}}{a numeric vector}
    \item{\code{x.19}}{a numeric vector}
    \item{\code{x.20}}{a numeric vector}
    \item{\code{x.21}}{a numeric vector}
  }
}
\examples{
str(waveform.test)
}
\keyword{datasets}
