\name{waveform.train}
\alias{waveform.train}
\docType{data}
\title{ Simulated Waveform Data}
\description{
  This is a simulated three-class problem with 21 variables, and 
  considered to be a difficult pattern recognition problem.  
}
\usage{data(waveform.train)}
\format{
  A data frame with 300 observations on the following 22 variables.
  \describe{
    \item{y}{a numeric vector}
    \item{x.1}{a numeric vector}
    \item{x.2}{a numeric vector}
    \item{x.3}{a numeric vector}
    \item{x.4}{a numeric vector}
    \item{x.5}{a numeric vector}
    \item{x.6}{a numeric vector}
    \item{x.7}{a numeric vector}
    \item{x.8}{a numeric vector}
    \item{x.9}{a numeric vector}
    \item{x.10}{a numeric vector}
    \item{x.11}{a numeric vector}
    \item{x.12}{a numeric vector}
    \item{x.13}{a numeric vector}
    \item{x.14}{a numeric vector}
    \item{x.15}{a numeric vector}
    \item{x.16}{a numeric vector}
    \item{x.17}{a numeric vector}
    \item{x.18}{a numeric vector}
    \item{x.19}{a numeric vector}
    \item{x.20}{a numeric vector}
    \item{x.21}{a numeric vector}
  }
}
\examples{
str(waveform.train)
}
\keyword{datasets}
